requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.8
httpx==0.24.1
anyio==3.7.1
sounddevice==0.4.6
numpy==1.24.3
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-xdist==3.5.0
